move moderateImage to utils
