+++ apps/src/lab2/utils/index.ts
@@ -7,3 +7,4 @@ export * from './fetchPermissions';
 export * from './getLabViewPageAction';
 export * from './getIsLabViewBlocked';
 export * from './convertProjectTypeToDisplayName';
+export * from './moderateImage';
